# docs/swagger.yaml
definitions:
...
  meetings.CreateMeetingInput:
    properties:
      date:
        description: Формат "YYYY-MM-DD"
        type: string
      end_time:
        description: Формат "HH:MM"
        type: string
      meeting_type:
        description: '"online" или "offline"'
        type: string
      room:
        description: Обязательное для офлайн встреч
        type: string
      start_time:
        description: Формат "HH:MM"
        type: string
      title:
        type: string
    required:
    - date
    - end_time
    - meeting_type
    - start_time
    - title
    type: object
  meetings.TimeSlot:
    properties:
      end:
        description: Например, "13:20"
        type: string
      start:
        description: Например, "12:00"
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  response.MeetingResponse:
    properties:
      conference_link:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      date:
        type: string
      end_time:
        type: string
      id:
        type: integer
      meeting_type:
        type: string
      room:
        type: string
      start_time:
        type: string
      team_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  title: Сервис для контроля задачами и встречами команды
paths:
...
  /meetings:
    post:
      consumes:
      - application/json
      description: Создает новую встречу для команды. Доступно только для менеджеров.
      parameters:
      - description: Уникальный идентификатор Telegram
        in: query
        name: telegram_id
        required: true
        type: string
      - description: Данные встречи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/meetings.CreateMeetingInput'
      produces:
      - application/json
      responses:
        "200":
          description: Информация о созданной встрече
          schema:
            $ref: '#/definitions/response.MeetingResponse'
        "400":
          description: Ошибка валидации или некорректные данные
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Доступ запрещен (не менеджер)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Конфликт по времени и аудитории
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Создание встречи
      tags:
      - meetings
  /meetings/available-slots:
    get:
      consumes:
      - application/json
      description: Возвращает список доступных временных слотов для указанной аудитории
        на выбранную дату
      parameters:
      - description: Номер аудитории
        in: query
        name: room
        required: true
        type: string
      - description: Дата в формате YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список доступных временных слотов
          schema:
            items:
              items:
                $ref: '#/definitions/meetings.TimeSlot'
              type: array
            type: array
        "400":
          description: Отсутствуют обязательные параметры или неверный формат
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка при получении данных
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получение доступных временных слотов
      tags:
      - meetings
...
swagger: "2.0"